Include error logs in email
Include error logs when sending email notifications of a submission error. The
admin email includes the full error log and the user email includes a minimal
error log. The minimal error log contains the specific error from the scoring
container to help the user fix their submission.

Additionally, the admin email template for a successful submission is updated
for consistency with the error template with respect to the included user
information.

Fixes #168