Send email notifications to challenge/phase admins instead of site admins
Previously, email notifications of successful submissions and submission errors
were sent to site admins. This change sends the notifications to challenge/phase
admins instead. In this context, an admin is a user with WRITE access or above
on the challenge or phase.

Email notifications to the user remain unchanged.

Fixes #162