Add basic submission dashboard
This adds an administrator interface to view all submissions to a phase. A
button at the bottom of the phase page routes to this interface.

Currently, the view shows submissions grouped by participant. By default,
submissions for up to five participants are shown at a time. The controls at the
top of the interface page through the list. The 'participantLimit' query
parameter sets the number of participants per page.

By default, up to five submissions per participant are shown at a time. The
controls in the group for each user page through the list. The 'submissionLimit'
query parameter sets the number of submissions to show per page.

Each submission shows its title, whether it's the latest successful submission,
its job status, its creation date, and its overall score. The job status text
links to the job detail page. Unsuccessful jobs include a tooltip showing the
first few lines of the job log. The overall score text links to the score detail
page. Links are provided to download the submission data and to view the raw
submission JSON.

To provide some extra information, warning messages are shown when a user hasn't
submitted at all or when the current page shows no successful submissions.

Some useful extensions, perhaps requiring backend enhancements to be efficient,
to this could include:
- display success/error/total submission counts per user/phase/challenge
- identify and group participants who have no successful submissions
- identify and group participants who haven't submitted
- filtering/sorting tables