Support viewing submission JSON for private phases
Previously, the link to view submission JSON used the covalic_submission/:id
endpoint. This would fail if the phase was not public, because that endpoint
does not support retrieving the token from a cookie.

This commit changes the link to instead dump the submission JSON to the console.
This is possible because the view already has the submission model. Dumping to
the console seems satisfactory, because one reason to view the submission JSON
is to be able to easily copy/paste IDs.