Fix job log emails
Fix a problem where the job log emails don't include the submission job
log. The emails erroneously stated that "No log is available."

The first problem is that the log sent in the jobs.job.update event is
no longer necessarily up-to-date, and may be empty. According to the
documentation [1], the job document should be manually re-refetched to
get the complete log.

The second problem is that the log is now a list, and the list elements
aren't necessarily broken at newlines. To fix this, we join the list
into a string before processing it.

[1] http://girder.readthedocs.io/en/latest/plugins.html