Add phase view to configure submissions
Add a new view to configure submissions for a phase. The new view
appears as the final step in the phase creation wizard and is accessible
later from the phase menu.

The "Require submission filenames to match ground truth filenames"
option is moved from the new phase/edit phase views to the configure
submissions view.