Ensure that metric popovers display at right position
When the score table is wide enough to require horizontal scrolling, the
metric popovers displayed when hovering over the column headers were not
displayed at the correct position for columns outside of the initial
viewport.

Specifying the popover 'viewport' property fixes the positioning.

Note that for very wide tables, the popover still doesn't appear in the
correct position, perhaps because of some internal limit in Bootstrap or
another component.