Enable re-qualifying a submission
Allow re-qualifying a submission so that it again appears in the
leaderboard. Note that the endpoint does not enforce the requiremenet
that only a single submission per user per phase be marked as the
'latest' submission. If access to the endpoint is expand beyond admin
users, then that requirement should be enforced.