Add multiple approaches to the submission model
This adds the ability for users to submit multiple approaches for each
phase of the challenge.  The approach is stored as a property on the
submission model.  The approach is expected to be an arbitrary string
that is coerced to lower case.

By default, a special approach called "default" is used.  To maintain
backward compatibility with existing databases, the default approach is
stored in the database by the absence of the "approach" key.  The save
and load methods are overridden to handle this behavior at the model
layer.

This also adds a new endpoint to list existing approaches by user and/or
phase.  The list is returned in alphebetical order and "default" is
always injected.

Future work could include adding settings to limit the number of
approaches allowed (even hard code a list of valid approaches) as part
of the model validation.