Alias empty string to the default approach
This change modifies the model layer so that the default approach can be
referenced as either 'default' or ''.  When saving to the database, the
validation method deletes the approach key if either of those values is
provided.

An equivalent change is made to the listing method.  A query on approach
with either a value of '' or 'default' will filter submissions in the
database with an approach of null.